+++ CHANGELOG.md
@@ -7,6 +7,9 @@ project adheres to [Semantic Versioning](http://semver.org/).
 - Added SHA256 checksum calculation to rake_helper.rb.
 - Added :use_random_org key to Passphrase#inspect.
 
+### Changed
+- Made :number_of_words an attribute of the Passphrase class.
+
 ## Version 1.0.0 - 2015-04-06
 This release is a complete re-write. None of the code from the previous
 release was retained.